alef: add meter settings migration and refactor entities

+++ migrations/007.do.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS payments (
     invoice_id INTEGER NOT NULL REFERENCES invoices(id),
     payment_date DATE NOT NULL,
     payment_hash UUID UNIQUE NOT NULL,
-    payment_amount INTEGER NOT NULL,
+    payment_amount DECIMAL(10, 2) NOT NULL,
     payment_method TEXT NOT NULL,
     payment_status TEXT NOT NULL,
     created_at TIMESTAMP NOT NULL DEFAULT NOW(),
